100% coverage with "real" tests, on both real and fake time
