+++ common/clock/ratelimiter_test.go
@@ -24,521 +24,444 @@ package clock
 
 import (
 	"context"
-	"fmt"
-	"math/rand"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/atomic"
 	"golang.org/x/sync/errgroup"
 	"golang.org/x/time/rate"
 )
 
-const (
-	// amount of time between each conceptual "tick" of the test's clocks, both real and fake.
-	// sadly even 10ms has some excessive latency a few % of the time, particularly during the Wait check.
-	granularity = 20 * time.Millisecond
+func TestRatelimiter(t *testing.T) {
+	t.Parallel()
 
-	// keep running tests until this amount of time has passed, or a failure has occurred.
-	// this could be a static count of tests to run, but a target duration
-	// is a bit less fiddly.  "many attempts" is the goal, not any specific number.
-	//
-	// alternatively, tests have a deadline, this could run until near that deadline.
-	// but right now that's not fine-tuned per package, so it's quite long.
-	fuzzDuration = 4 * time.Second // mostly stays under 5s, feels reasonable
-
-	// debug-level output is very noisy on successful runs, so hide it by default.
-	// using a custom seed will also set this to true.
-	enableDebugLog = false
-	customSeed     = 0 // override with non-zero seed to run just one test with that seed
-)
-
-func TestAgainstRealRatelimit(t *testing.T) {
-	const (
-		// number of [granularity] events to trigger per round,
-		// and also the number of [granularity] periods before a token is added.
-		//
-		// each test takes at least granularity*events*rounds, so keep it kinda small.
-		events = 10
-		// number of rounds of "N events" to try
-		rounds = 2
-
-		// less than events to give Wait a better chance of waiting,
-		// but anything up to (events*rounds)/2 will have some effect.
-		maxBurst = events / 2
-	)
-
-	// The mock ratelimiter should behave the same as a real ratelimiter in all cases.
-	//
-	// So fuzz test it: throw random calls at non-overlapping times that are a bit
-	// spaced out (to prevent noise due to busy CPUs) and make sure all impls agree.
-	//
-	// If a test fails, please verify by hand with the failing seed.
-	// Enabling debug logs can help show what's happening and what *should* be happening,
-	// though there is a chance it's just due to CPU contention.
-	t.Run("fuzz", func(t *testing.T) {
-		deadline := time.Now().Add(fuzzDuration)
-		for testnum := 0; !t.Failed() && time.Now().Before(deadline); testnum++ {
-			if testnum > 0 && customSeed != 0 {
-				// already ran the seeded test, stop early.
-				break
+	for _, name := range []string{"real", "mocked"} {
+		name := name
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			ts := func() MockedTimeSource { return nil }
+			if name == "mocked" {
+				ts = NewMockedTimeSource
 			}
-			t.Run(fmt.Sprintf("attempt %v", testnum), func(t *testing.T) {
-				t.Cleanup(func() {
-					if t.Failed() {
-						/*
-							Is this test becoming too slow or too noisy?
+			// needs to be a bit coarser than the comparison tests, because time
+			// is less tightly controlled / mock time relies on a background loop.
+			granularity := 50 * time.Millisecond
+			assertRatelimiterBasicsWork(t, ts, granularity)
+		})
+	}
+}
 
-							The easiest "likely to actually work" fix is probably going to
-							require detecting excessive lag, and retrying instead of failing.
+// check relatively basic properties of the ratelimiter wrapper,
+// and make sure that it behaves the same with both real time (makeTimesource returns nil)
+// and with mocked time (simulated by advancing mock-time in a background goroutine).
+//
+// broken out because it's just very deeply indented otherwise.
+func assertRatelimiterBasicsWork(t *testing.T, makeTimesource func() MockedTimeSource, granularity time.Duration) {
+	// all "events" take place at 2*granularity
+	event := granularity * 2
 
-							That could miss some real flaws if they are racy by nature, but
-							seems like it might be good enough.
-						*/
-						t.Logf("---- CAUTION ----")
-						t.Logf("these tests are randomized by design, a random failure may be a real flaw!")
-						t.Logf("please replay with the failing seed (set `customSeed`) and check detailed output to see what the behavior should be.")
-						t.Logf("also try setting `enableDebugLog` to false to true to show Tokens() progression, this can help explain many issues.")
-						t.Logf("")
-						t.Logf("if you are intentionally making changes, be sure to run a few hundred rounds to make sure your changes are stable,")
-						t.Logf("and in particular make sure an 'Advancing mock time ...' event occurs and times match, as that is a bit rare")
-						t.Logf("---- CAUTION ----")
-					}
-				})
+	makeLimiter := func(ts MockedTimeSource, limit rate.Limit, burst int) (rl Ratelimiter, sleep func(time.Duration)) {
+		if ts == nil {
+			return NewRatelimiter(limit, burst), time.Sleep
+		}
+		return NewMockRatelimiter(ts, limit, burst), ts.Advance
+	}
 
-				// parallel saves a fair bit of time but introduces a lot of CPU contention
-				// and that leads to a moderate amount of flakiness.
-				//
-				// unfortunately not recommended here.
-				// t.Parallel()
+	now := func(ts MockedTimeSource) time.Time {
+		if ts == nil {
+			return time.Now()
+		}
+		return ts.Now()
+	}
 
-				seed := time.Now().UnixNano()
-				if customSeed != 0 {
-					seed = customSeed // override seed to test a specific scenario
+	// creates a context that will time out with the passed timesource, or real time if nil
+	contextWithTimeout := func(ts TimeSource, dur time.Duration) (ctx context.Context, cancel context.CancelFunc) {
+		if ts == nil {
+			// could be a timesourceContext with a real timesource,
+			// but I'm avoiding that until it's fleshed out fully and tested separately.
+			return context.WithTimeout(context.Background(), dur)
+		}
+		ctx = &timesourceContext{
+			ts:       ts,
+			deadline: ts.Now().Add(dur),
+		}
+		return ctx, func() {}
+	}
+
+	// advances mock time in the background until stopped, at [granularity/10] per millisecond tick.
+	// currently this leads to tests that are about 5x faster than real time, but "faster" is not
+	// a goal in this context.
+	//
+	// "different" is however good for making sure we don't rely on real time somehow,
+	// so try to keep it either faster or slower, whatever makes for a decently non-flaky suite.
+	//
+	// flawed ratelimiter behavior when using mock time should also cause the fuzz test to fail
+	// eventually, so failures here are, hopefully, only due to flawed tests or excessive noise.
+	advanceTime := func(ts MockedTimeSource) (cleanup func()) {
+		if ts == nil {
+			return func() {} // nothing to do
+		}
+		stop, done := make(chan struct{}), make(chan struct{})
+		go func() {
+			defer close(done)
+			// this is not a great way to do mock-time tests, but it does automatically
+			// work in just about all cases, without needing customization.
+			//
+			// ideally, in other kinds of tests, you should wait for ts.BlockUntil(..)
+			// or some other event, and then advance time semi-precisely, and fall back
+			// to this strategy only when that isn't usable for some reason.
+			//
+			// one extremely nice quality of mocked time though, even when done like this:
+			// time simply pauses while debugging, and nothing "artificially" times out
+			// due to being blocked, regardless of how long you let things sit.
+			ticker := time.NewTicker(time.Millisecond)
+			defer ticker.Stop()
+			for {
+				select {
+				case <-stop:
+					return
+				case <-ticker.C:
+					ts.Advance(granularity / 10)
 				}
-				rng := rand.New(rand.NewSource(seed))
-				t.Logf("rand seed: %v", seed)
-				round := makeRandomSchedule(t, rng, rounds, events)
-				burst := rng.Intn(maxBurst) // zero is not very interesting, but allowed
-				// waits do not usually need to pause and it's tough to assert randomly.
-				// any waited periods are still asserted to be similar though.
-				// the expected-wait case is tested separately below.
-				noMinWait := time.Duration(0)
-				assertRatelimitersBehaveSimilarly(t, burst, noMinWait, round)
+			}
+		}()
+		return func() {
+			close(stop)
+			<-done
+		}
+	}
+
+	t.Run("sets burst and limit", func(t *testing.T) {
+		t.Parallel()
+		// should set on construction
+		rl, _ := makeLimiter(makeTimesource(), rate.Every(time.Second), 3)
+		assert.EqualValues(t, 3, rl.Burst())
+		assert.Equal(t, rate.Every(time.Second), rl.Limit())
+
+		// and when calling the setters
+		rl.SetBurst(5)
+		rl.SetLimit(rate.Every(time.Millisecond))
+		assert.EqualValues(t, 5, rl.Burst())
+		assert.Equal(t, rate.Every(time.Millisecond), rl.Limit())
+	})
+	t.Run("limits requests to within available burst tokens", func(t *testing.T) {
+		t.Parallel()
+		rl, _ := makeLimiter(makeTimesource(), rate.Every(time.Second), 3)
+		assert.EqualValues(t, 3, rl.Tokens(), "should have tokens to allow exactly 3 calls, and not exceed burst of 3")
+		assert.True(t, rl.Allow())
+		assert.InDeltaf(t, 2, rl.Tokens(), 0.1, "Allow should consume a token")
+		assert.True(t, rl.Allow())
+		assert.InDelta(t, 1, rl.Tokens(), 0.1)
+		assert.True(t, rl.Allow())
+		assert.InDelta(t, 0, rl.Tokens(), 0.1)
+		assert.False(t, rl.Allow(), "should not allow after burstable tokens are consumed")
+		assert.InDelta(t, 0, rl.Tokens(), 0.1)
+		assert.False(t, rl.Allow())
+	})
+	t.Run("recovers tokens as time passes", func(t *testing.T) {
+		t.Parallel()
+		rl, sleep := makeLimiter(makeTimesource(), rate.Every(event), 1)
+		assert.True(t, rl.Allow())
+		assert.False(t, rl.Allow())
+		sleep(event)
+		assert.True(t, rl.Allow(), "should have recovered one token")
+		assert.False(t, rl.Allow())
+	})
+	t.Run("waits until tokens are available", func(t *testing.T) {
+		t.Parallel()
+		ts := makeTimesource()
+		rl, _ := makeLimiter(ts, rate.Every(event), 1)
+		started := now(ts) // must be collected before Allow, or sleep time could legitimately be lower than the rate
+		assert.True(t, rl.Allow())
+		assert.False(t, rl.Allow())
+
+		stopTime := advanceTime(ts)
+		ctx, cancel := contextWithTimeout(ts, event)
+		defer cancel()
+		err := rl.Wait(ctx)
+		stopTime()
+
+		assert.NoError(t, err, "Wait should have waited and then allowed the event")
+		elapsed := now(ts).Sub(started)
+		assert.Truef(t, elapsed >= event-granularity && elapsed < event+granularity,
+			"elapsed time outside bounds, should be %v <= %v < %v",
+			event-granularity, elapsed, event+granularity)
+
+		assert.False(t, rl.Allow(), "Wait must consume a token")
+	})
+	t.Run("reserve", func(t *testing.T) {
+		t.Parallel()
+		// simple stuff
+		for name, tc := range map[string]struct {
+			do          func(t *testing.T, rl Ratelimiter, sleep func(time.Duration))
+			tokenChange float64
+		}{
+			"consumes a token": {
+				do: func(t *testing.T, rl Ratelimiter, sleep func(time.Duration)) {
+					r := rl.Reserve()
+					assert.True(t, r.Allow())
+					r.Used(true)
+				},
+				tokenChange: -1,
+			},
+			"restores a token when canceling": {
+				do: func(t *testing.T, rl Ratelimiter, sleep func(time.Duration)) {
+					r := rl.Reserve()
+					assert.True(t, r.Allow())
+					r.Used(false)
+				},
+				tokenChange: 0,
+			},
+			"does not restore if there are interleaving calls": {
+				do: func(t *testing.T, rl Ratelimiter, sleep func(time.Duration)) {
+					r := rl.Reserve()
+					assert.True(t, r.Allow())
+					sleep(1) // advance time by any amount
+					rl.Allow()
+					r.Used(false)
+				},
+				tokenChange: -2, // reservation did not restore back to -1
+			},
+			"does not restore if Tokens is called in before canceling": {
+				do: func(t *testing.T, rl Ratelimiter, sleep func(time.Duration)) {
+					// to make a possibly-surprising quirk explicit:
+					// tokens advances the limiter's "now" value, which means
+					// canceling an allowed token does not work.
+					r := rl.Reserve()
+					assert.True(t, r.Allow())
+					sleep(1)    // advance time by any amount
+					rl.Tokens() // advance time internally
+					r.Used(false)
+				},
+				tokenChange: -1, // -1 due to Tokens call
+			},
+			"does not go negative": {
+				do: func(t *testing.T, rl Ratelimiter, sleep func(time.Duration)) {
+					rl.Allow()
+					rl.Allow()
+					assert.InDeltaf(t, 0, rl.Tokens(), 0.1, "should have consumed all tokens")
+					r := rl.Reserve()
+					assert.False(t, r.Allow())
+					r.Used(false)
+				},
+				tokenChange: -2, // would be -3 if it went negative
+			},
+		} {
+			name, tc := name, tc
+			t.Run(name, func(t *testing.T) {
+				t.Parallel()
+				ts := makeTimesource()
+				rl, sleep := makeLimiter(ts, rate.Every(event), 2)
+
+				before := rl.Tokens()
+				stopTime := advanceTime(ts)
+				tc.do(t, rl, sleep)
+				stopTime()
+				after := rl.Tokens()
+
+				assert.InDeltaf(t, tc.tokenChange, after-before, 0.1, "tokens should have changed from %v to %v, but was %v", before, before+tc.tokenChange, after)
 			})
 		}
 	})
-
-	// also check an edge case that random tests only rarely hit:
-	// Wait at the end of a round until a token recovers at the start of the next.
-	t.Run("wait_must_wait_similarly", func(t *testing.T) {
-		schedule := [][]string{
-			{"a", "_", "_", "w"}, // consume the only token, then wait with 0.75 tokens.
-			{"_", "_", "_", "_"}, // first slot must be _ to recover a token and unblock the wait.
+	t.Run("wait", func(t *testing.T) {
+		t.Parallel()
+		wait := func(ts MockedTimeSource, periods float64, rl Ratelimiter) int64 {
+			ctx, cancel := contextWithTimeout(ts, time.Duration(periods*float64(event)))
+			defer cancel()
+			if rl.Wait(ctx) == nil {
+				return 1
+			}
+			return 0
 		}
-		logSchedule(t, schedule)
-		minWait := granularity / 2 // not waiting will fail this test
-		burst := 1
-		assertRatelimitersBehaveSimilarly(t, burst, minWait, schedule)
-	})
-}
-
-/*
-Generate a random schedule for testing behavior.
-
-Schedules look like:
-
-	[r _ a w _ _ _ _ _ w]
-	[_ c _ _ _ _ w _ _ _]
-
-Which means that the first round will:
-  - reserve()
-  - do nothing
-  - allow()
-  - wait() (may extend to next event, so the next one is always "_")
-  - ...
-
-By the end of that first array it'll reach the end of the `rate.Every` time, and
-may recover tokens during the next round's runtime.  So the second round will:
-  - do nothing, just refresh .1 while waiting ==> will have 1.0 tokens regenerated
-    since the test began, as it was reserved earlier (if non-zero burst).
-    (this could be immediately consumed by last round's final wait)
-  - reserve but cancel, restoring the token (if any), and maybe refresh .1 token
-  - do nothing, maybe refresh .1 more token...
-  - ... and try to wait on the 7th event.
-
-Non-blank events do not overlap between rounds to avoid triggering a race
-between the old consumed tokens being refreshed and a new call consuming them,
-as these make for very flaky tests unless time is mocked.
-*/
-func makeRandomSchedule(t *testing.T, rng *rand.Rand, rounds, events int) [][]string {
-	// generate some non-colliding "A == perform an Allow call" rounds.
-	schedule := make([][]string, rounds) // using string so it prints nicely
-	for i := 0; i < rounds; i++ {
-		round := make([]string, events)
-		for j := 0; j < events; j++ {
-			round[j] = "_" // "do nothing" marker
-		}
-		schedule[i] = round
-	}
-	skipWait := false
-	for i := 0; i < len(schedule[0]); i++ {
-		if skipWait {
-			skipWait = false
-			continue
-		}
-		// pick one to set to true, or none (1 chance for none)
-		set := rng.Intn(len(schedule) + 1)
-		if set < len(schedule) {
-			// 1/3rd of them are waits
-			switch rng.Intn(4) {
-			case 0:
-				schedule[set][i] = "a" // allow
-			case 1:
-				schedule[set][i] = "r" // reserve
-			case 2:
-				schedule[set][i] = "c" // reserve but cancel
-			case 3:
-				schedule[set][i] = "w" // wait
-				// waits can pause for an event and a half, so skip the next event to avoid overlapping
-				skipWait = true
-			default:
-				panic("missing a case")
+		parWait := func(ts MockedTimeSource, periods float64, count *atomic.Int64, rl Ratelimiter) {
+			if wait(ts, periods, rl) == 1 {
+				count.Inc()
 			}
 		}
-	}
-	logSchedule(t, schedule)
-
-	return schedule
-}
-
-func logSchedule(t *testing.T, schedule [][]string) {
-	t.Log("round setup:")
-	for _, round := range schedule {
-		t.Logf("\t%v", round)
-	}
-}
-
-func debugLogf(t *testing.T, format string, args ...interface{}) {
-	// by default be silent, but log if checking a specific seed or force-enabled.
-	if enableDebugLog || customSeed != 0 {
-		t.Helper()
-		t.Logf(format, args...)
-	}
-}
-
-func assertRatelimitersBehaveSimilarly(t *testing.T, burst int, minWaitDuration time.Duration, schedule [][]string) {
-	rounds, eventsPerRound := len(schedule), len(schedule[0])
-	limit := rate.Every(granularity * time.Duration(eventsPerRound)) // refresh after each full round
-	t.Logf("limit: %v, burst: %v", granularity, burst)
-
-	initialNow := time.Now().Round(time.Millisecond) // easier to read when logging/debugging when rounded
-	ts := NewMockedTimeSourceAt(initialNow)
-	compressedTS := NewMockedTimeSourceAt(initialNow)
-
-	actual := rate.NewLimiter(limit, burst)
-	wrapped := NewRatelimiter(limit, burst)
-	mocked := NewMockRatelimiter(ts, limit, burst)
-	compressed := NewMockRatelimiter(compressedTS, limit, burst)
-
-	// record "to be executed" closures on the time-compressed ratelimiter too,
-	// so it can be checked at a sped up rate.
-	// nil values mean "nothing to do".
-	compressedReplay := make([][]func(t *testing.T), rounds)
-	for i := range compressedReplay {
-		compressedReplay[i] = make([]func(t *testing.T), eventsPerRound)
-	}
-
-	ticker := time.NewTicker(granularity)
-	defer ticker.Stop()
-	for round := range schedule {
-		round := round // for closure
-		for event := range schedule[round] {
-			event := event // for closure
-			<-ticker.C
-			ts.Advance(granularity) // may also be advanced inside wait
-			debugLogf(t, "Tokens before round, real: %0.2f, wrapped: %0.2f, mocked: %0.2f", actual.Tokens(), wrapped.Tokens(), mocked.Tokens())
-
-			switch schedule[round][event] {
-			case "a":
-				// call Allow on everything
-				a, w, m := actual.Allow(), wrapped.Allow(), mocked.Allow()
-				assertAllowsAgree(t, "Allow", round, event, a, w, m, nil)
-				compressedReplay[round][event] = func(t *testing.T) {
-					c := compressed.Allow()
-					assertAllowsAgree(t, "Allow", round, event, a, w, m, &c)
-				}
-			case "r":
-				// call Reserve on everything, don't cancel any we acquired successfully
-				now := time.Now()
-				_a, _w, _m := actual.ReserveN(now, 1), wrapped.Reserve(), mocked.Reserve()
-				a, w, m := _a.OK() && _a.DelayFrom(now) == 0, _w.Allow(), _m.Allow()
-
-				// un-reserve any that were not successful, we don't leave them dangling.
-				if !a {
-					_a.CancelAt(now)
-				}
-				_w.Used(w)
-				_m.Used(m)
-				assertAllowsAgree(t, "Reserve", round, event, a, w, m, nil)
-				compressedReplay[round][event] = func(t *testing.T) {
-					_c := compressed.Reserve()
-					c := _c.Allow()
-					_c.Used(c)
-					assertAllowsAgree(t, "Reserve", round, event, a, w, m, &c)
-				}
-			case "c":
-				// call Reserve on everything, and cancel them all
-				now := time.Now() // needed for the real ratelimiter to cancel successfully like the wrapper does
-				_a, _w, _m := actual.ReserveN(now, 1), wrapped.Reserve(), mocked.Reserve()
-				_a.CancelAt(now)
-				_w.Used(false)
-				_m.Used(false)
-				t.Logf("round[%v][%v] Cancel:     canceled", round, event)
-				compressedReplay[round][event] = func(t *testing.T) {
-					compressed.Reserve().Used(false)
-					t.Logf("round[%v][%v] Cancel:     canceled", round, event)
-				}
-			case "w":
-				/*
-					Try a brief Wait on everything.
-
-					ctx should expire in the middle of the *next* event, so all three possibilities are exercised:
-					  - sometimes unblocks immediately because a token is available
-					  - sometimes waits because a token will become available soon
-					  - sometimes unblocks immediately because a token will not become available in time
-
-					you can see waits logged like this, they're somewhat rare though.  otherwise wait times are ~0s:
-						regular:
-							ratelimiter_test.go:402: Wait elapsed: 20ms, wrapped err: <nil>
-							ratelimiter_test.go:639: Advancing mock time by 20ms
-							ratelimiter_test.go:402: Wait elapsed: 20ms, mocked err: <nil>
-							ratelimiter_test.go:402: Wait elapsed: 20ms, actual err: <nil>
-							ratelimiter_test.go:655: round[0][9] Wait:       actual: true,  wrapped: true,  mocked: true,
-							ratelimiter_test.go:455:             Wait time:  actual: 20ms,  wrapped: 20ms,  mocked: 20ms,
-						compressed:
-							ratelimiter_test.go:672: Advancing mock time by 20ms
-							ratelimiter_test.go:402: Wait elapsed: 20ms, compressed err: <nil>
-							ratelimiter_test.go:684: round[0][9] Wait:       actual: true,  wrapped: true,  mocked: true,  compressed: true
-							ratelimiter_test.go:455:             Wait time:  actual: 20ms,  wrapped: 20ms,  mocked: 20ms,  compressed: 20ms
-				*/
-				timeout := granularity + (granularity / 2)
-				started := time.Now() // intentionally gathered outside the goroutines, to reveal goroutine-starting lag
-				ctx, cancel := context.WithTimeout(context.Background(), timeout)
-
-				a, w, m := false, false, false
-				var aLatency, wLatency, mLatency time.Duration
-				var g errgroup.Group
-				g.Go(func() error {
-					_a := actual.Wait(ctx)
-					aLatency = time.Since(started).Round(time.Millisecond)
-					debugLogf(t, "Wait elapsed: %v, actual err: %v", aLatency, _a)
-					a = _a == nil
-					return nil
-				})
-				g.Go(func() error {
-					_w := wrapped.Wait(ctx)
-					wLatency = time.Since(started).Round(time.Millisecond)
-					debugLogf(t, "Wait elapsed: %v, wrapped err: %v", wLatency, _w)
-					w = _w == nil
-					return nil
-				})
-				g.Go(func() error {
-					// mocked time needs something to advance it, or it'll block until timeout...
-					// but don't advance time if it returned immediately, the others won't wait either
-					done := make(chan struct{})
-					go func() {
-						select {
-						case <-time.After(granularity):
-							t.Logf("Advancing mock time by %v", granularity)
-							ts.Advance(granularity)
-						case <-done:
-							// do nothing, it is not waiting
-						}
-					}()
-
-					_m := mocked.Wait(ctx)
-					close(done)
-					mLatency = time.Since(started).Round(time.Millisecond)
-					debugLogf(t, "Wait elapsed: %v, mocked err: %v", mLatency, _m)
-					m = _m == nil
-					return nil
-				})
-				_ = g.Wait()
-
-				assertAllowsAgree(t, "Wait", round, event, a, w, m, nil)
-				assertLatenciesSimilar(t, minWaitDuration, timeout, aLatency, wLatency, mLatency, nil)
-				compressedReplay[round][event] = func(t *testing.T) {
-					// need a mocked-time context, or the real deadline will not match the mocked deadline
-					ctx := &timesourceContext{
-						ts:       compressedTS,
-						deadline: compressedTS.Now().Add(granularity + (granularity / 2)),
+		for name, tc := range map[string]struct {
+			drainFirst  bool
+			do          func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64
+			latency     float64 // num of [granularity]s that should elapse
+			allowed     int64
+			tokenChange float64
+		}{
+			"allows immediately with free tokens": {
+				drainFirst: false,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					return wait(ts, 3, rl) // anything longer than 1
+				},
+				allowed:     1,
+				tokenChange: -1,
+			},
+			"fails immediately with too-short timeout": {
+				drainFirst: true,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					return wait(ts, 0.1, rl)
+				},
+				latency: 0,
+				allowed: 0,
+			},
+			"waits for one token": {
+				drainFirst: true,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					return wait(ts, 2, rl)
+				},
+				latency: 1,
+				allowed: 1,
+			},
+			"concurrently waits for 2 tokens": {
+				drainFirst: true,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					count := atomic.Int64{}
+					var g errgroup.Group
+					for i := 0; i < 2; i++ {
+						g.Go(func() error {
+							parWait(ts, 2.5, &count, rl) // slightly longer than 2
+							return nil
+						})
 					}
-
-					done := make(chan struct{})
+					g.Wait()
+					return count.Load()
+				},
+				latency: 2,
+				allowed: 2,
+			},
+			"one wins and one fails when concurrently waiting": {
+				drainFirst: true,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					count := atomic.Int64{}
+					var g errgroup.Group
+					for i := 0; i < 2; i++ {
+						g.Go(func() error {
+							parWait(ts, 1.5, &count, rl) // slightly longer than 1
+							return nil
+						})
+					}
+					g.Wait()
+					// only one can win
+					return count.Load()
+				},
+				latency: 1,
+				allowed: 1,
+			},
+			"immediately drains 2, waits for 2, fails 2": {
+				drainFirst: false,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					count := atomic.Int64{}
+					var g errgroup.Group
+					for i := 0; i < 6; i++ {
+						g.Go(func() error {
+							parWait(ts, 2.5, &count, rl) // slightly longer than 2, so 2 waiters succeed
+							return nil
+						})
+					}
+					g.Wait()
+					return count.Load()
+				},
+				latency:     2,
+				allowed:     2,
+				tokenChange: -2,
+			},
+			"cancel while waiting": {
+				drainFirst: true,
+				do: func(t *testing.T, rl Ratelimiter, ts MockedTimeSource) int64 {
+					var wg sync.WaitGroup
+					count := atomic.Int64{}
+					wg.Add(2)
+					ctx, cancel := context.WithTimeout(context.Background(), event*3/2)
 					go func() {
-						// wait until blocked inside Wait
-						compressedTS.BlockUntil(1)
-						// sleep briefly to encourage close(done) if racing, or blocked on something else
-						time.Sleep(granularity / 10)
-						select {
-						case <-done:
-							// do nothing, Wait did not block correctly.
-							// this goroutine may leak if no blockers occur. it's "fine".
-						default:
-							// blocked inside Wait, advance time so it unblocks
-							t.Logf("Advancing mock time by %v", granularity)
-							compressedTS.Advance(granularity)
+						defer wg.Done()
+						// a bit too much of a mess to thread this through nicely
+						if ts == nil {
+							<-time.After(event / 2) // cancel half way through the wait
+							cancel()
+						} else {
+							ts.AfterFunc(event/2, func() {
+								cancel()
+							})
 						}
 					}()
-
-					compressedStarted := compressedTS.Now()
-					_c := compressed.Wait(ctx)
-					close(done)
-					cLatency := time.Since(started).Round(time.Millisecond)
-					cLatency = compressedTS.Since(compressedStarted)
-					debugLogf(t, "Wait elapsed: %v, compressed err: %v", cLatency, _c)
-					c := _c == nil
-					assertAllowsAgree(t, "Wait", round, event, a, w, m, &c)
-					assertLatenciesSimilar(t, minWaitDuration, timeout, aLatency, wLatency, mLatency, &cLatency)
+					go func() {
+						defer wg.Done()
+						if rl.Wait(ctx) == nil {
+							count.Inc()
+						}
+					}()
+					wg.Wait()
+					return count.Load()
+				},
+				allowed:     0,
+				latency:     0.5, // unfortunately allows 0 currently, but token change should ensure time passed
+				tokenChange: 0.5, // only waited for half an event -> recovered half a token
+			},
+		} {
+			name, tc := name, tc
+			t.Run(name, func(t *testing.T) {
+				t.Parallel()
+				ts := makeTimesource()
+				rl, _ := makeLimiter(ts, rate.Every(granularity*2), 2)
+				if tc.drainFirst {
+					assert.True(t, rl.Allow())
+					assert.True(t, rl.Allow())
+					assert.False(t, rl.Allow())
 				}
-				cancel()
-			case "_":
-				// do nothing
-			default:
-				panic("unhandled case: " + schedule[round][event])
-			}
 
-			debugLogf(t, "Tokens after round, real: %0.2f, wrapped: %0.2f, mocked: %0.2f", actual.Tokens(), wrapped.Tokens(), mocked.Tokens())
-		}
-	}
-	t.Run("compressed time", func(t *testing.T) {
-		// and now replay the compressed ratelimiter and make sure it matches too,
-		// as time-compressed must behave the same as real-time.
-		//
-		// this is primarily intended to detect cases where real-time is accidentally used,
-		// as ~zero time actually passes, which is quite different from the above tests.
-		//
-		// it's not perfect, but it does eventually notice such bugs, as you can see by
-		// changing literally any timesource.Now() calls into time.Now() and running
-		// tests a few times.
-		// depending on the change it might not notice in most tests, but eventually a
-		// problematic combination is triggered and can be replayed with the logged seed.
-		for round := range schedule {
-			for event := range schedule[round] {
-				compressedTS.Advance(granularity)
-				debugLogf(t, "Tokens before compressed round: %0.2f", compressed.Tokens())
-				replay := compressedReplay[round][event]
-				if replay != nil {
-					replay(t)
-				}
-				debugLogf(t, "Tokens after compressed round: %0.2f", compressed.Tokens())
-			}
+				before := rl.Tokens()
+				start := now(ts)
+				stopTime := advanceTime(ts)
+				tc.do(t, rl, ts)
+				stopTime()
+				after := rl.Tokens()
+				elapsed := now(ts).Sub(start)
+				target := time.Duration(tc.latency * float64(granularity*2))
+
+				assert.InDeltaf(t, tc.tokenChange, after-before, 0.1, "tokens should have changed from %0.2f to %0.2f, but was %0.2f -> %0.2f", before, before+tc.tokenChange, before, after)
+				assert.True(t,
+					elapsed > (target-granularity) && elapsed < (target+granularity),
+					"should have taken %v < (actual) %v < %v",
+					target-granularity, elapsed, target+granularity,
+				)
+			})
 		}
 	})
 }
 
-// broken out to consts to help keep padding the same, for easier visual scanning
-const (
-	similarBehaviorLogComponents  = "%-11s %-14s %-15s %-14s" // "{what:} {actual} {wrapped} {mocked}" with padding
-	similarBehaviorLogNormalRound = "round[%v][%v] " + similarBehaviorLogComponents
-)
-
-func assertAllowsAgree(t *testing.T, what string, round, event int, actual, wrapped, mocked bool, compressed *bool) {
-	t.Helper()
-	sWhat := what + ":"
-	sActual := fmt.Sprintf("actual: %v,", actual)
-	sWrapped := fmt.Sprintf("wrapped: %v,", wrapped)
-	sMocked := fmt.Sprintf("mocked: %v,", mocked)
-	// format strings should share padding so output aligns and it's easy to read.
-	// unfortunately only strings do this nicely.s
-	if compressed != nil {
-		sCompressed := "compressed: " + fmt.Sprint(*compressed)
-		t.Logf(similarBehaviorLogNormalRound+" %s", round, event, sWhat, sActual, sWrapped, sMocked, sCompressed)
-		assert.True(t, actual == wrapped && wrapped == mocked && mocked == *compressed, "ratelimiters disagree")
-	} else {
-		t.Logf(similarBehaviorLogNormalRound, round, event, sWhat, sActual, sWrapped, sMocked)
-		assert.True(t, actual == wrapped && wrapped == mocked, "ratelimiters disagree")
-	}
+func TestRatelimiterCoverage(t *testing.T) {
+	t.Run("time paradox", func(t *testing.T) {
+		// this test's behavior should not be possible, as any reservation with
+		// a time in the future must have been reserved in the future, which
+		// would have advanced the internal latestNow to match that future-time.
+		rl := NewRatelimiter(rate.Every(time.Second), 1)
+		impl := rl.(*ratelimiter)
+		r := rl.Reserve()
+		latestNow := impl.latestNow
+		require.True(t, r.Allow())
+		rimpl := r.(*allowedReservation)
+		// cheat, put the reservation into the future.
+		rimpl.reservedAt = time.Now().Add(time.Second)
+		rimpl.Used(false)
+		assert.True(t, latestNow.Before(impl.latestNow), "ratelimiter-internal time should have advanced despite impossible sequence")
+	})
+	t.Run("zero burst", func(t *testing.T) {
+		// we do not currently make use of this, but it's allowed by the API.
+		rl := NewRatelimiter(rate.Every(time.Second), 0)
+		ctx, cancel := context.WithTimeout(context.Background(), time.Hour) // much longer than the rate
+		defer cancel()
+		started := time.Now()
+		assert.Error(t, rl.Wait(ctx), "0 burst should never allow events, so it should fail immediately")
+		elapsed := time.Since(started)
+		assert.Less(t, elapsed, time.Second/10, "Wait should have returned almost immediately on impossible waits")
+	})
+	t.Run("mock limiter constructor", func(t *testing.T) {
+		// covered by fuzz testing, but this gets it to 100% without fuzz.
+		_ = NewMockRatelimiter(NewMockedTimeSource(), 1, 1)
+	})
 }
 
-func assertLatenciesSimilar(t *testing.T, mustBeGreaterThan, mustBeLessThan, actual, wrapped, mocked time.Duration, compressed *time.Duration) {
-	t.Helper()
-	// none are currently expected to wait, but if they do, they must not wait the full timeout.
-	// this also helps handle small cpu stutter, as some may wait 1ms or so.
-	maxLatency := maxDur(actual, wrapped, mocked)
-	minLatency := minDur(actual, wrapped, mocked)
-	compressedString := ""
-	if compressed != nil {
-		maxLatency = maxDur(maxLatency, *compressed)
-		minLatency = minDur(minLatency, *compressed)
-		compressedString = fmt.Sprintf("compressed: %v", *compressed)
-	}
-	assert.GreaterOrEqualf(t, minLatency, mustBeGreaterThan, "minimum latency must be within bounds, else waits did not wait long enough")
-	assert.LessOrEqualf(t, maxLatency, mustBeLessThan, "maximum latency must not exceed timeout, it implies wait did not unblock when it should")
-
-	// log format tries to align with assertAllowsAgree
-	indent := strings.Repeat(" ", len("round[1][1]"))
-	sActual := fmt.Sprintf("actual: %v,", actual)
-	sWrapped := fmt.Sprintf("wrapped: %v,", wrapped)
-	sMocked := fmt.Sprintf("mocked: %v,", mocked)
-	t.Logf("%s "+similarBehaviorLogComponents+" %s",
-		indent, "Wait time:", sActual, sWrapped, sMocked, compressedString)
-
-	// asserting is... tough.  long-ish noise latencies are somewhat common,
-	// sometimes even exceeding 10ms to schedule a goroutine.
-	//
-	// if this becomes too noisy, just switch to logging, and check by hand
-	// when making changes.
-	if minLatency < granularity/2 {
-		// "no wait" test, assert that they are all under this granularity
-		assertNoWait := func(what string, wait time.Duration) {
-			t.Helper()
-			assert.LessOrEqual(t, wait, granularity/2, "%v ratelimiter waited too long", what)
-		}
-		assertNoWait("actual", actual)
-		assertNoWait("wrapped", wrapped)
-		assertNoWait("mocked", mocked)
-		if compressed != nil {
-			assertNoWait("compressed", *compressed)
-		}
-	} else {
-		// "wait" test, assert that they all waited, and none waited significantly longer
-		assertWaited := func(what string, wait time.Duration) {
-			t.Helper()
-			assert.True(t,
-				wait > granularity/2 && wait < granularity+(granularity/2),
-				"%v waited an incorrect amount of time, should be %v < %v <%v",
-				what, granularity/2, wait, granularity+(granularity/2))
-		}
-		assertWaited("actual", actual)
-		assertWaited("wrapped", wrapped)
-		assertWaited("mocked", mocked)
-		if compressed != nil {
-			assertWaited("compressed", *compressed)
-		}
-	}
-}
-
-func maxDur(d time.Duration, ds ...time.Duration) time.Duration {
-	for _, tmp := range ds {
-		if tmp > d {
-			d = tmp
-		}
-	}
-	return d
-}
-func minDur(d time.Duration, ds ...time.Duration) time.Duration {
-	for _, tmp := range ds {
-		if tmp < d {
-			d = tmp
-		}
-	}
-	return d
-}
-
+// context which becomes Done() based on a TimeSource instead of real time
 type timesourceContext struct {
 	// does not contain a parent context as we currently have no need,
 	// but a "real" one would for forwarding Value and deadline lookups.
